+++ Tests/KeyboardKitTests/Mocks/MockKeyboardContext.swift
@@ -16,9 +16,6 @@ class MockKeyboardContext: KeyboardContext {
     var device: UIDevice = .current
     
     var actionHandler: KeyboardActionHandler = MockKeyboardActionHandler()
-    var keyboardInputSetProvider: KeyboardInputSetProvider = StaticKeyboardInputSetProvider.empty
-    var keyboardLayoutProvider: KeyboardLayoutProvider = StandardKeyboardLayoutProvider()
-    var secondaryCalloutActionProvider: SecondaryCalloutActionProvider = StandardSecondaryCalloutActionProvider()
     
     var deviceOrientation: UIInterfaceOrientation = .portrait
     var emojiCategory: EmojiCategory = .frequent