+++ RELEASE_NOTES.md
@@ -19,6 +19,8 @@ Although these release notes will aim at covering everything that changes in thi
 * `KeyboardInputViewController` has a new `keyboardActionHandler`.
 * `KeyboardInputViewController` has a new `keyboardAppearance`.
 * `KeyboardInputViewController` has a new `keyboardBehavior`.
+* `KeyboardInputViewController` has a new `keyboardContext`.
+* `KeyboardInputViewController` has a new `keyboardLayoutProvider`.
 * `View` has new `keyboardInputViewController` convenience extensions.